📌 ISSUE-#30: Code improvement
