+++ dotflow/__init__.py
@@ -4,7 +4,7 @@ __version__ = "0.4.0"
 
 from .core.action import Action as action
 from .core.context import Context
-from .core.workflow import DotFlow
+from .core.dotflow import DotFlow
 from .core.task import Task
 from .core.decorators import retry  # deprecated
