+++ dotflow/core/models/__init__.py
@@ -1,10 +1,10 @@
 """Types __init__ module."""
 
-from dotflow.core.models.execution import Execution
+from dotflow.core.models.execution import TypeExecution
 from dotflow.core.models.status import Status
 
 
 __all__ = [
-    "Execution",
+    "TypeExecution",
     "Status"
 ]
