+++ dotflow/core/task.py
@@ -13,28 +13,22 @@ from dotflow.core.utils import callback
 class Task:
 
     def __init__(
-            self,
-            task_id: int,
-            step: Callable,
-            callback: Callable = callback,
-            initial_context: Any = None,
-            current_context: Any = None,
-            previous_context: Any = None,
-            status: Status = Status.NOT_STARTED,
-            error: List[Exception] = [],
-            duration: float = 0,
-            workflow_id: UUID = None
+        self,
+        task_id: int,
+        step: Callable,
+        callback: Callable = callback,
+        initial_context: Any = None,
     ) -> None:
         self.task_id = task_id
         self.step = step
         self.callback = callback
         self.initial_context = Context(initial_context)
-        self.current_context = Context(current_context)
-        self.previous_context = Context(previous_context)
-        self.status = status
-        self.error = error
-        self.duration = duration
-        self.workflow_id = workflow_id
+        self.current_context = Context()
+        self.previous_context = Context()
+        self.status = Status.NOT_STARTED
+        self.error = []
+        self.duration = 0
+        self.workflow_id = None
 
     def set_status(self, value: Status) -> None:
         self.status = value
@@ -48,6 +42,9 @@ class Task:
     def set_previous_context(self, value: Context) -> None:
         self.previous_context = value
 
+    def set_workflow_id(self, value: UUID) -> None:
+        self.workflow_id = value
+
 
 class TaskBuilder:
 
@@ -55,10 +52,7 @@ class TaskBuilder:
         self.queu: List[Task] = []
 
     def add(
-            self,
-            step: Callable,
-            callback: Callable = callback,
-            initial_context: Any = None
+        self, step: Callable, callback: Callable = callback, initial_context: Any = None
     ) -> None:
         if step.__module__ != Action.__module__:
             raise MissingActionDecorator()
